+++ Editor/Drawers/DisabledDrawers/ReadOnlyAttributeDrawer.cs
@@ -103,7 +103,7 @@ namespace SaintsField.Editor.Drawers.DisabledDrawers
 
             foreach (string andCallback in bys ?? Array.Empty<string>())
             {
-                (string error, bool isTruly) = IsTruly(target, type, andCallback);
+                (string error, bool isTruly) = Util.GetTruly(target, type, andCallback);
                 if (error != "")
                 {
                     errors.Add(error);
@@ -123,58 +123,6 @@ namespace SaintsField.Editor.Drawers.DisabledDrawers
             Debug.Log($"final return={truly}/editor[{editorMode}]={editorModeIsTrue}/bys={string.Join(",", callbackTruly)}");
 #endif
             return ("", truly);
-        }
-
-        protected static (string error, bool isTruly) IsTruly(object target, Type type, string by)
-        {
-            (ReflectUtils.GetPropType getPropType, object fieldOrMethodInfo) = ReflectUtils.GetProp(type, by);
-
-            if (getPropType == ReflectUtils.GetPropType.NotFound)
-            {
-                string error = $"No field or method named `{by}` found on `{target}`";
-                // Debug.LogError(error);
-                // _errors.Add(error);
-                return (error, false);
-            }
-
-            if (getPropType == ReflectUtils.GetPropType.Property)
-            {
-                return ("", ReflectUtils.Truly(((PropertyInfo)fieldOrMethodInfo).GetValue(target)));
-            }
-            if (getPropType == ReflectUtils.GetPropType.Field)
-            {
-                return ("", ReflectUtils.Truly(((FieldInfo)fieldOrMethodInfo).GetValue(target)));
-            }
-            // ReSharper disable once InvertIf
-            if (getPropType == ReflectUtils.GetPropType.Method)
-            {
-                MethodInfo methodInfo = (MethodInfo)fieldOrMethodInfo;
-                ParameterInfo[] methodParams = methodInfo.GetParameters();
-                Debug.Assert(methodParams.All(p => p.IsOptional));
-                object methodResult;
-                // try
-                // {
-                //     methodInfo.Invoke(target, methodParams.Select(p => p.DefaultValue).ToArray())
-                // }
-                try
-                {
-                    methodResult = methodInfo.Invoke(target, methodParams.Select(p => p.DefaultValue).ToArray());
-                }
-                catch (TargetInvocationException e)
-                {
-                    Debug.LogException(e);
-                    Debug.Assert(e.InnerException != null);
-                    return (e.InnerException.Message, false);
-
-                }
-                catch (Exception e)
-                {
-                    Debug.LogException(e);
-                    return (e.Message, false);
-                }
-                return ("", ReflectUtils.Truly(methodResult));
-            }
-            throw new ArgumentOutOfRangeException(nameof(getPropType), getPropType, null);
         }
         #endregion