+++ Editor/Playa/Renderer/NonSerializedFieldRenderer.cs
@@ -1,4 +1,6 @@
-﻿using SaintsField.Editor.Core;
+﻿using System.Linq;
+using SaintsField.Editor.Core;
+using SaintsField.Playa;
 using UnityEditor;
 using UnityEngine;
 #if UNITY_2022_2_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
@@ -17,12 +19,25 @@ namespace SaintsField.Editor.Playa.Renderer
         public override VisualElement CreateVisualElement()
         {
             object value = FieldWithInfo.FieldInfo.GetValue(SerializedObject.targetObject);
-            return UIToolkitLayout(value, ObjectNames.NicifyVariableName(FieldWithInfo
+            VisualElement result = UIToolkitLayout(value, ObjectNames.NicifyVariableName(FieldWithInfo
                 .FieldInfo.Name));
+
+            if (FieldWithInfo.PlayaAttributes.Count(each => each is PlayaShowIfAttribute || each is PlayaHideIfAttribute) > 0)
+            {
+                result.RegisterCallback<AttachToPanelEvent>(_ => result.schedule.Execute(() => UIToolkitOnUpdate(result, false)).Every(100));
+            }
+
+            return result;
         }
 #endif
         public override void Render()
         {
+            PreCheckResult preCheckResult = GetPreCheckResult(FieldWithInfo);
+            if (!preCheckResult.IsShown)
+            {
+                return;
+            }
+
             object value = FieldWithInfo.FieldInfo.GetValue(SerializedObject.targetObject);
             FieldLayout(value, ObjectNames.NicifyVariableName(FieldWithInfo
                 .FieldInfo.Name));
@@ -30,11 +45,23 @@ namespace SaintsField.Editor.Playa.Renderer
 
         public override float GetHeight()
         {
+            PreCheckResult preCheckResult = GetPreCheckResult(FieldWithInfo);
+            if (!preCheckResult.IsShown)
+            {
+                return 0;
+            }
+
             return SaintsPropertyDrawer.SingleLineHeight;
         }
 
         public override void RenderPosition(Rect position)
         {
+            PreCheckResult preCheckResult = GetPreCheckResult(FieldWithInfo);
+            if (!preCheckResult.IsShown)
+            {
+                return;
+            }
+
             object value = FieldWithInfo.FieldInfo.GetValue(FieldWithInfo.Target);
             FieldPosition(position, value, ObjectNames.NicifyVariableName(FieldWithInfo
                 .FieldInfo.Name));