+++ Editor/Utils/Util.cs
@@ -41,50 +41,6 @@ namespace SaintsField.Editor.Utils
 
         public static (string error, float value) GetCallbackFloat(object target, string by)
         {
-            // SerializedProperty foundProperty = property.FindPropertyRelative(by) ??
-            //                                    SerializedUtils.FindPropertyByAutoPropertyName(property.serializedObject, by);
-            // if (foundProperty != null)
-            // {
-            //     if (foundProperty.propertyType == SerializedPropertyType.Integer)
-            //     {
-            //         return (foundProperty.intValue, null);
-            //     }
-            //     if (foundProperty.propertyType == SerializedPropertyType.Float)
-            //     {
-            //         return (foundProperty.floatValue, null);
-            //     }
-            //
-            //     return (-1, $"Expect int or float for `{by}`, get {foundProperty.propertyType}");
-            // }
-            //
-            // object target = property.serializedObject.targetObject;
-
-            // (ReflectUtils.GetPropType getPropType, object fieldOrMethodInfo) found = ReflectUtils.GetProp(target.GetType(), by);
-            // switch (found)
-            // {
-            //     case (ReflectUtils.GetPropType.NotFound, _):
-            //     {
-            //         return (-1, $"No field or method named `{by}` found on `{target}`");
-            //     }
-            //     case (ReflectUtils.GetPropType.Property, PropertyInfo propertyInfo):
-            //     {
-            //         return ObjToFloat(propertyInfo.GetValue(target));
-            //     }
-            //     case (ReflectUtils.GetPropType.Field, FieldInfo foundFieldInfo):
-            //     {
-            //         return ObjToFloat(foundFieldInfo.GetValue(target));
-            //     }
-            //     case (ReflectUtils.GetPropType.Method, MethodInfo methodInfo):
-            //     {
-            //         ParameterInfo[] methodParams = methodInfo.GetParameters();
-            //         Debug.Assert(methodParams.All(p => p.IsOptional));
-            //         // Debug.Assert(methodInfo.ReturnType == typeof(bool));
-            //         return ObjToFloat(methodInfo.Invoke(target, methodParams.Select(p => p.DefaultValue).ToArray()));
-            //     }
-            //     default:
-            //         throw new ArgumentOutOfRangeException(nameof(found), found, null);
-            // }
-
             (ReflectUtils.GetPropType getPropType, object fieldOrMethodInfo) found = ReflectUtils.GetProp(target.GetType(), by);
 
             // ReSharper disable once ConvertIfStatementToSwitchStatement
@@ -334,6 +290,58 @@ namespace SaintsField.Editor.Utils
             }
 
             return false;
+        }
+
+        public static (string error, bool isTruly) GetTruly(object target, Type type, string by)
+        {
+            (ReflectUtils.GetPropType getPropType, object fieldOrMethodInfo) = ReflectUtils.GetProp(type, by);
+
+            if (getPropType == ReflectUtils.GetPropType.NotFound)
+            {
+                string error = $"No field or method named `{by}` found on `{target}`";
+                // Debug.LogError(error);
+                // _errors.Add(error);
+                return (error, false);
+            }
+
+            if (getPropType == ReflectUtils.GetPropType.Property)
+            {
+                return ("", ReflectUtils.Truly(((PropertyInfo)fieldOrMethodInfo).GetValue(target)));
+            }
+            if (getPropType == ReflectUtils.GetPropType.Field)
+            {
+                return ("", ReflectUtils.Truly(((FieldInfo)fieldOrMethodInfo).GetValue(target)));
+            }
+            // ReSharper disable once InvertIf
+            if (getPropType == ReflectUtils.GetPropType.Method)
+            {
+                MethodInfo methodInfo = (MethodInfo)fieldOrMethodInfo;
+                ParameterInfo[] methodParams = methodInfo.GetParameters();
+                Debug.Assert(methodParams.All(p => p.IsOptional));
+                object methodResult;
+                // try
+                // {
+                //     methodInfo.Invoke(target, methodParams.Select(p => p.DefaultValue).ToArray())
+                // }
+                try
+                {
+                    methodResult = methodInfo.Invoke(target, methodParams.Select(p => p.DefaultValue).ToArray());
+                }
+                catch (TargetInvocationException e)
+                {
+                    Debug.LogException(e);
+                    Debug.Assert(e.InnerException != null);
+                    return (e.InnerException.Message, false);
+
+                }
+                catch (Exception e)
+                {
+                    Debug.LogException(e);
+                    return (e.Message, false);
+                }
+                return ("", ReflectUtils.Truly(methodResult));
+            }
+            throw new ArgumentOutOfRangeException(nameof(getPropType), getPropType, null);
         }
     }
 }