+++ Runtime/Playa/OrderedAttribute.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 namespace SaintsField.Playa
 {
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Method | AttributeTargets.Property)]
-    public class OrderedAttribute: Attribute
+    public class OrderedAttribute: Attribute, IPlayaAttribute
     {
         public readonly int Order;