+++ Runtime/Playa/ShowInInspectorAttribute.cs
@@ -3,7 +3,7 @@
 namespace SaintsField.Playa
 {
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Method | AttributeTargets.Property)]
-    public class ShowInInspectorAttribute: Attribute
+    public class ShowInInspectorAttribute: Attribute, IPlayaAttribute
     {
     }
 }